Moved existing tests for DropDown from Repository tests to Shared view tests.
